Fix current warnings, better OSC Config errors, debug/release conf, etc
