+++ src/errors.rs
@@ -15,11 +15,11 @@ pub enum AppError {
         path: PathBuf,
         source: std::io::Error,
     },
-    #[error("Failed to write to file \"{path}\": {source}")]
-    WriteFile {
-        path: PathBuf,
-        source: std::io::Error,
-    },
+    // #[error("Failed to write to file \"{path}\": {source}")]
+    // WriteFile {
+    //     path: PathBuf,
+    //     source: std::io::Error,
+    // },
     #[error("IO error: {0}")]
     Io(#[from] std::io::Error),
     #[error("Error parsing IP Address: {0}")]
@@ -34,8 +34,13 @@ pub enum AppError {
     Bt(#[from] btleplug::Error),
     #[error("TOML Serialization Error: {0}")]
     TomlSer(#[from] toml::ser::Error),
-    #[error("Failed to get working directory")]
-    WorkDir,
     #[error("CSV Error: {0}")]
     Csv(#[from] csv_async::Error),
+    // My errors
+    #[error("Failed to get working directory")]
+    WorkDir,
+    #[error("Invalid OSC Prefix: \"{0}\"")]
+    OscPrefix(String),
+    #[error("Invalid OSC Address: \"{0}\" - \"{1}\"")]
+    OscAddress(String, String),
 }
