+++ src/logging.rs
@@ -6,9 +6,7 @@ use crate::settings::MiscSettings;
 use csv_async::AsyncSerializer;
 use log::*;
 use serde::Serialize;
-use std::env;
 use std::path::PathBuf;
-use std::sync::Arc;
 use std::time::Duration;
 use tokio::fs::{create_dir, File};
 use tokio::io::{AsyncSeekExt, AsyncWriteExt, BufWriter};
